pyproject.toml
[tool.black]
skip-string-normalization = true

[tool.poetry]
name = "yearn-exporter"
version = "1.0.0"
description = "yearn exporter"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
eth-brownie = "==1.19.3"
click = ">=7.1.2"
tabulate = ">=0.8.7"
joblib = ">=1.0.1"
cachetools = ">=4.1.1"
fastapi = ">=0.63.0"
uvicorn = ">=0.13.4"
sqlmodel = ">=0.0.4"
psycopg2-binary = ">=2.8.5"
#tokenlists = ">=0.1.7,<0.2"
semantic-version = ">=2.8.5"
boto3 = ">=1.34,<1.35"
matplotlib = ">=3.6.1"
sqlalchemy = ">=1.4.41"
sentry-sdk = ">=2.1.1,<2.2"
pytest-bdd = ">=5.0.0"
dank_mids = "4.20.94"
eth_retry = ">=0.1.19"
eth-hash = { version = "*", extras = ["pysha3"] }
ez-a-sync = "0.22.3"
#ez-a-sync = { git = "https://www.github.com/BobTheBuidler/ez-a-sync", branch = "smarter" }
python-telegram-bot = "13.15"
ypricemagic = { git = "https://www.github.com/BobTheBuidler/ypricemagic", branch = "master" }
#ypricemagic = { git = "https://www.github.com/BobTheBuidler/ypricemagic", branch = "temp-fix" }
#eth-portfolio = { git = "https://www.github.com/BobTheBuidler/eth-portfolio", branch = "ypm-3.6.1" }
eth-portfolio = { git = "https://www.github.com/BobTheBuidler/eth-portfolio", branch = "faster-restarts" }
toolcache = { git = "https://www.github.com/BobTheBuidler/toolcache", rev = "e37b53cec64556d2b35df66767fca4c8f366a2fa" }
memray = "*"
numpy = "<2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
